refactor: update FormConfigComponent tests to use mockResponse and improve structure
